Module Scad_ml

module Scad : sig ... end

The core module of this library. Containing the core type, and all operations upon it.

module Text : sig ... end

Types pertaining to the configuration of OpenSCAD text objects.

module Color : sig ... end

Type defining the available colour options in OpenSCAD.

module MultMatrix : sig ... end

An affine transformation matrix.

module RotMatrix : sig ... end

A rotation matrix.

module Quaternion : sig ... end

Provides functions for the creation of and operations between quaternions. These can be used to create composable and interpolatable rotations to be applied to vectors (e.g. Vec3.t) directly, and Scad.t through MultMatrix.t.

module Vec3 : sig ... end

3-dimensional vector type.