Module Scad_ml__Scad

The core module of this library. Containing the core type, and all operations upon it.

OpenSCAD objects and the operations upon them are represented by the recursive type t, which can then finally be translated into a scad script. As this transpiles down to OpenSCAD, the User Manual can be referred to for additional information about most of the functions made available here.

type t

3d shape primitives

val cube : ?⁠center:bool -> (float * float * float) -> t

cube ?center dimensions

val sphere : ?⁠fa:float -> ?⁠fs:float -> ?⁠fn:int -> float -> t

sphere ?fa ?fs ?fn radius

val cylinder : ?⁠center:bool -> ?⁠fa:float -> ?⁠fs:float -> ?⁠fn:int -> float -> float -> t

cylinder ?center ?fa ?fs ?fn radius height

val polyhedron : Scad_ml.Vec3.t list -> int list list -> t

polyhedron points faces

2d shape primitives

val square : ?⁠center:bool -> (float * float) -> t

square ?center dimensions

val circle : ?⁠fa:float -> ?⁠fs:float -> ?⁠fn:int -> float -> t

circle ?fa ?fs ?fn radius

val polygon : ?⁠convexity:int -> ?⁠paths:int list list -> (float * float) list -> t

polygon ?convexity ?paths points

val text : ?⁠size:float -> ?⁠font:string -> ?⁠halign:Scad_ml.Text.h_align -> ?⁠valign:Scad_ml.Text.v_align -> ?⁠spacing:float -> ?⁠direction:Scad_ml.Text.direction -> ?⁠language:string -> ?⁠script:string -> ?⁠fn:int -> string -> t

text ?size ?font ?halign ?valign ?spacing ?direction ?language ?script ?fn str

Transformations

val translate : Scad_ml.Vec3.t -> t -> t

translate p t

val rotate : Scad_ml.Vec3.t -> t -> t

rotate r t

val rotate_about_pt : Scad_ml.Vec3.t -> Scad_ml.Vec3.t -> t -> t

rotate_about_pt r p t

val vector_rotate : Scad_ml.Vec3.t -> float -> t -> t

vector_rotate ax r t

val vector_rotate_about_pt : Scad_ml.Vec3.t -> float -> Scad_ml.Vec3.t -> t -> t

vector_rotate_about_pt ax r p t

val multmatrix : Scad_ml.MultMatrix.t -> t -> t

multmatrix mat t

val mirror : (float * float * float) -> t -> t

mirror ax t

val quaternion : Scad_ml.Quaternion.t -> t -> t

quaternion q t

val quaternion_about_pt : Scad_ml.Quaternion.t -> Scad_ml.Vec3.t -> t -> t

quaternion_about_pt q p t

val scale : (float * float * float) -> t -> t

scale factors t

val resize : (float * float * float) -> t -> t

resize dimensions t

val offset : ?⁠chamfer:bool -> [ `Delta of float | `Radius of float ] -> t -> t

offset ?chamfer offset t

val color : ?⁠alpha:float -> Scad_ml.Color.t -> t -> t

color ?alpha color t

Boolean Combination

val union : t list -> t

union ts

val minkowski : t list -> t

minkowski ts

val hull : t list -> t

hull ts

val difference : t -> t list -> t

difference t sub

val intersection : t list -> t

intersection ts

3d to 2d

val projection : ?⁠cut:bool -> t -> t

projection ?cut t

2d to 3d extrusions

val linear_extrude : ?⁠height:float -> ?⁠center:bool -> ?⁠convexity:int -> ?⁠twist:int -> ?⁠slices:int -> ?⁠scale:float -> ?⁠fn:int -> t -> t

linear_extrude ?height ?center ?convexity ?twist ?slices ?scale ?fn t

val rotate_extrude : ?⁠angle:float -> ?⁠convexity:int -> ?⁠fa:float -> ?⁠fs:float -> ?⁠fn:int -> t -> t

rotate_extrude ?angle ?convexity ?fa ?fs ?fn t

val import : ?⁠dxf_layer:string -> ?⁠convexity:int -> string -> t

import ?dxf_layer ?convexity file

val (|>>) : t -> Scad_ml.Vec3.t -> t

t |>> p

Infix translate

val (|@>) : t -> Scad_ml.Vec3.t -> t

t |@> r

Infix rotate

val to_string : t -> string

to_string t

val write : Stdlib.out_channel -> t -> unit

write oc t